+++ docs/examples/docs-demo-six-datepicker-13.vue
@@ -1,17 +1,38 @@
 <template>
 <div>
 
-        <six-datepicker id="allowed-date-picker"></six-datepicker>
+        <div class="min-max-demo">
+          <six-datepicker id="min-max-date-picker"></six-datepicker>
+          <six-datepicker id="only-min-date-picker"></six-datepicker>
+          <six-datepicker id="only-max-date-picker"></six-datepicker>
+        </div>
+
+        <div style="margin-top: 1rem">
+          <span style="font-weight: bold">Selected: </span><span id="min-max-date-selected">No Date selected yet!</span>
+          
+        </div>
 
         
+
         
       
 </div>
 </template>
 <style>
 
-          six-datepicker {
-            max-width: 25rem;
+            six-datepicker {
+              max-width: 25rem;
+            }
+          
+          .min-max-demo {
+            display: flex;
+            flex-wrap: wrap;
+          }
+
+          .min-max-demo six-datepicker {
+            min-width: 15rem;
+            margin-right: 1rem;
+            margin-bottom: 1rem;
           }
         
 </style>
@@ -19,8 +40,23 @@
 export default {
   name: 'docs-demo-six-datepicker-13',
   mounted() { 
-          const datepicker = document.getElementById('allowed-date-picker');
-          datepicker.allowedDates = (date) => date.getDate() % 2 === 0;
+          const minMaxDatepicker = document.getElementById('min-max-date-picker');
+          const minDatepicker = document.getElementById('only-min-date-picker');
+          const maxDatepicker = document.getElementById('only-max-date-picker');
+          const selectedDate = document.getElementById('min-max-date-selected');
+
+          const getDateFromNow = (days) => new Date(new Date().setDate(new Date().getDate() + days));
+
+          minMaxDatepicker.value = new Date();
+          minMaxDatepicker.min = getDateFromNow(-3);
+          minMaxDatepicker.max = getDateFromNow(2);
+
+          minDatepicker.min = getDateFromNow(-3);
+          maxDatepicker.max = getDateFromNow(2);
+
+          minMaxDatepicker.addEventListener('six-datepicker-select', (event) => {
+            selectedDate.innerHTML = `${event.detail.toLocaleDateString()}`;
+          });
          }
 }
 </script>
