+++ docs/examples/docs-demo-six-datepicker-17.vue
@@ -1,7 +1,11 @@
 <template>
 <div>
 
-        <six-datepicker custom-icon>
+        <six-datepicker icon-position="right"></six-datepicker>
+        <br>
+        <six-datepicker icon-position="right" clearable></six-datepicker>
+        <br>
+        <six-datepicker icon-position="right" custom-icon clearable>
           <img             style="height: 24px"
             slot="custom-icon"
             src="https://upload.wikimedia.org/wikipedia/commons/thumb/5/59/Sketch_Logo.svg/1133px-Sketch_Logo.svg.png"
