+++ docs/examples/docs-demo-six-datepicker-26.vue
@@ -1,0 +1,21 @@
+<template>
+<div>
+
+        <six-datepicker invalid>
+          <div slot="error-text">
+            <six-error               >An error message
+              <a href="https://github.com/six-group/six-webcomponents" target="_blank">with a link</a></six-error>
+          </div>
+        </six-datepicker>
+      
+</div>
+</template>
+<style>
+
+</style>
+<script>
+export default {
+  name: 'docs-demo-six-datepicker-26',
+  mounted() {  }
+}
+</script>
