+++ docs/examples/docs-demo-six-datepicker-6.vue
@@ -1,7 +1,7 @@
 <template>
 <div>
 
-        <six-datepicker placement="bottom"></six-datepicker>
+        <six-datepicker readonly></six-datepicker>
         
       
 </div>
